changing some stuff
